Remove changes related to pod watch
