+++ pkg/controller/v1beta1/inferenceservice/suite_test.go
@@ -98,15 +98,11 @@ var _ = BeforeSuite(func() {
 	Expect(err).ToNot(HaveOccurred())
 	Expect(clientset).ToNot(BeNil())
 
-	cacheOpts, err := NewCacheOptions()
-	Expect(err).ToNot(HaveOccurred())
-
 	k8sManager, err := ctrl.NewManager(cfg, ctrl.Options{
 		Scheme: scheme.Scheme,
 		Metrics: metricsserver.Options{
 			BindAddress: "0",
 		},
-		Cache: cacheOpts,
 	})
 	Expect(err).ToNot(HaveOccurred())
